// src/app/request/interfaces/request.interfaces.ts
export interface request {
  type_edification?:    string;
  areas_demolishions?:  number;
  type_material?:       string;
  electric_pole?:       string;
  value_Demolishion?:   number;
  number_levels?:       number;
  type_Energy?:         number;
  Contruccion_disable?: number;
  address?:             string;
  contruction_company?: number;
  user_id?:             number;
  paz_save?:File;
  approval?:File;
  status?:string;
  thecnical_resolution?:File;
  updated_at?:          Date;
  created_at?:          Date;
  id?:                  number;
}